requirements.txt
telebot
requests
pyodbc
python-dotenv
telethon
asyncio